+++ packages/block-library/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add background color controls for the table block.
 - Add new `RSS` block ([#7966](https://github.com/WordPress/gutenberg/pull/7966)).
+- Add new `Search` block ([#13583](https://github.com/WordPress/gutenberg/pull/13583)).
 
 ## 2.2.12 (2019-01-03)